requirements.txt
google-generativeai
ddgs
rich
pandas
requests
beautifulsoup4
python-dotenv
